Move to utils/pasting.js, merge w/ get-paste-event-data.js
